script to verify and fix bad LFS lazy copies
not sure where the bug is, but for some reason the LFS lazy copies are copied from the wrong file revisions if there are more than one source revision. this results in errors when pulling from lfs.

the script in question will verify the lfs objects in the .git-fusion depot, identify bad ones, locate the correct revision of the source file, and replace the lfs objects with fixed ones using p4 copy. p4gf_clear_cache.py needs to be run on the git-fusion server as the git user to make sure the updates come through